Split tasks into toplevel.
