+++ tasks/classifyapp/dataset/poj104.cc
@@ -13,7 +13,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#include "programl/task/classifyapp/dataset/poj104.h"
+#include "tasks/classifyapp/dataset/poj104.h"
 
 #include <stdio.h>
