+++ tasks/dataflow/BUILD
@@ -84,10 +84,10 @@ py_binary(
         "//programl/proto:epoch_py",
         "//programl/proto:program_graph_features_py",
         "//programl/proto:program_graph_py",
-        "//programl/task/dataflow",
-        "//programl/task/dataflow:ggnn_batch_builder",
-        "//programl/task/dataflow:vocabulary",
-        "//programl/task/dataflow/dataset:pathflag",
+        "//tasks/dataflow",
+        "//tasks/dataflow:ggnn_batch_builder",
+        "//tasks/dataflow:vocabulary",
+        "//tasks/dataflow/dataset:pathflag",
         "//third_party/py/labm8",
         "//third_party/py/numpy",
     ],
@@ -144,7 +144,7 @@ py_binary(
         ":vocabulary",
         "//programl/proto:node_py",
         "//programl/proto:program_graph_py",
-        "//programl/task/dataflow/dataset:pathflag",
+        "//tasks/dataflow/dataset:pathflag",
         "//third_party/py/labm8",
         "//third_party/py/ncc:vocabulary",
     ],
