+++ tasks/dataflow/dataset/unpack_ir_lists.cc
@@ -28,7 +28,7 @@
 #include "labm8/cpp/logging.h"
 #include "labm8/cpp/strutil.h"
 #include "programl/proto/ir.pb.h"
-#include "programl/task/dataflow/dataset/parallel_file_map.h"
+#include "tasks/dataflow/dataset/parallel_file_map.h"
 
 namespace fs = boost::filesystem;
 using std::vector;
