+++ tasks/devmap/dataset/BUILD
@@ -1,0 +1,38 @@
+# Copyright 2019-2020 the ProGraML authors.
+#
+# Contact Chris Cummins <chrisc.101@gmail.com>.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+py_binary(
+    name = "create",
+    srcs = ["create.py"],
+    deps = [
+        "//programl/ir/llvm/py:llvm",
+        "//programl/proto:features_py",
+        "//tasks/dataflow/dataset:pathflag",
+        "//third_party/py/labm8",
+        "//third_party/py/numpy",
+        "//third_party/py/pandas",
+    ],
+)
+
+py_test(
+    name = "create_test",
+    timeout = "long",
+    srcs = ["create_test.py"],
+    deps = [
+        ":create",
+        "//third_party/py/labm8",
+    ],
+)
