+++ tasks/devmap/dataset/create_test.py
@@ -1,11 +1,3 @@
-# Flag files for configuring data flow model jobs.
-#
-# Example Usage:
-#
-#    bazel run -c opt //programl/task/dataflow:train_ggnn -- \
-#       --flagfile=/path/to/flagfile.txt \
-#       --analysis=reachability
-#
 # Copyright 2019-2020 the ProGraML authors.
 #
 # Contact Chris Cummins <chrisc.101@gmail.com>.
@@ -21,13 +13,25 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+"""Smoke test for //tasks/devmap/dataset:create."""
+from pathlib import Path
 
-filegroup(
-    name = "flagfiles",
-    srcs = [
-        "cdfg.txt",
-        "inst2vec.txt",
-        "programl.txt",
-    ],
-    visibility = ["//programl/task/dataflow:__subpackages__"],
-)
+from labm8.py import test
+
+from programl.task.devmap.dataset.create import create_devmap_dataset
+
+
+def test_create_devmap_dataset(tempdir: Path):
+    """Test dataset creation."""
+    create_devmap_dataset(tempdir)
+    assert (tempdir / "ir").is_dir()
+    assert (tempdir / "src").is_dir()
+    assert (tempdir / "graphs_amd").is_dir()
+    assert (tempdir / "graphs_nvidia").is_dir()
+
+    assert len(list((tempdir / "graphs_amd").iterdir())) == 680
+    assert len(list((tempdir / "graphs_nvidia").iterdir())) == 680
+
+
+if __name__ == "__main__":
+    test.Main()
